backend/requirements.txt
fastapi==0.65.2
pydantic>=1.2.0,<2.0.0
uvicorn==0.12.1
mysql-connector-python
requests==2.24.0
sqlalchemy < 1.4.0
bcrypt==4.0.1
pyjwt==1.7.1
sqlalchemy-utils==0.36.8
passlib
pymysql
python-multipart